requirements.txt
litellm
google-generativeai
kokoro-onnx
soundfile
edge-tts
networkx
pydantic
pyyaml
